Remove auto-update and keep manual notices
