--- test/auto-update-checker.test.ts
@@ -1,1147 +1,0 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { EventEmitter } from "node:events";
-
-vi.mock("node:fs", () => ({
-	chmodSync: vi.fn(),
-	readFileSync: vi.fn(),
-	writeFileSync: vi.fn(),
-	existsSync: vi.fn(),
-	mkdirSync: vi.fn(),
-	renameSync: vi.fn(),
-	rmdirSync: vi.fn(),
-	statSync: vi.fn(),
-	unlinkSync: vi.fn(),
-}));
-vi.mock("node:child_process", () => ({
-	spawn: vi.fn(),
-}));
-
-describe("auto-update-checker", () => {
-	let fs: typeof import("node:fs");
-	let childProcess: typeof import("node:child_process");
-	let checkForUpdates: typeof import("../lib/auto-update-checker.js").checkForUpdates;
-	let checkAndNotify: typeof import("../lib/auto-update-checker.js").checkAndNotify;
-	let clearUpdateCache: typeof import("../lib/auto-update-checker.js").clearUpdateCache;
-	let autoUpdateIfAvailable: typeof import("../lib/auto-update-checker.js").autoUpdateIfAvailable;
-	let isAutoUpdateEnabled: typeof import("../lib/auto-update-checker.js").isAutoUpdateEnabled;
-	let isUpdateablePackageInstall: typeof import("../lib/auto-update-checker.js").isUpdateablePackageInstall;
-	let resolvePackageRootFromModuleDir: typeof import("../lib/auto-update-checker.js").resolvePackageRootFromModuleDir;
-	let logger: {
-		debug: ReturnType<typeof vi.fn>;
-		info: ReturnType<typeof vi.fn>;
-		warn: ReturnType<typeof vi.fn>;
-	};
-
-	const mockPackageJson = { name: "codex-multi-auth", version: "4.12.0" };
-
-	beforeEach(async () => {
-		vi.resetModules();
-		vi.useFakeTimers();
-		vi.setSystemTime(new Date("2026-01-30T12:00:00Z"));
-		mockPackageJson.version = "4.12.0";
-		logger = {
-			debug: vi.fn(),
-			info: vi.fn(),
-			warn: vi.fn(),
-		};
-		vi.doMock("../lib/logger.js", () => ({
-			createLogger: () => logger,
-		}));
-
-		fs = await import("node:fs");
-		childProcess = await import("node:child_process");
-		vi.mocked(childProcess.spawn).mockReset();
-		vi.mocked(fs.mkdirSync).mockReset();
-		vi.mocked(fs.renameSync).mockReset();
-		vi.mocked(fs.rmdirSync).mockReset();
-		vi.mocked(fs.statSync).mockReset();
-		vi.mocked(fs.unlinkSync).mockReset();
-		vi.mocked(fs.chmodSync).mockReset();
-		vi.mocked(fs.writeFileSync).mockReset();
-		vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-			if (String(path).includes("package.json")) {
-				return JSON.stringify(mockPackageJson);
-			}
-			throw new Error("File not found");
-		});
-		vi.mocked(fs.existsSync).mockReturnValue(false);
-		vi.mocked(fs.statSync).mockReturnValue({ mtimeMs: Date.now() } as never);
-
-		globalThis.fetch = vi.fn();
-
-		const module = await import("../lib/auto-update-checker.js");
-		checkForUpdates = module.checkForUpdates;
-		checkAndNotify = module.checkAndNotify;
-		clearUpdateCache = module.clearUpdateCache;
-		autoUpdateIfAvailable = module.autoUpdateIfAvailable;
-		isAutoUpdateEnabled = module.isAutoUpdateEnabled;
-		isUpdateablePackageInstall = module.isUpdateablePackageInstall;
-		resolvePackageRootFromModuleDir = module.resolvePackageRootFromModuleDir;
-	});
-
-	afterEach(() => {
-		vi.useRealTimers();
-		vi.restoreAllMocks();
-	});
-
-	describe("compareVersions (tested via checkForUpdates)", () => {
-		it("returns hasUpdate=true when latest > current", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(true);
-			expect(result.currentVersion).toBe("4.12.0");
-			expect(result.latestVersion).toBe("5.0.0");
-		});
-
-		it("returns hasUpdate=false when latest < current", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "3.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("returns hasUpdate=false when versions are equal", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "4.12.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("handles versions with v-prefix", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "v4.13.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("treats stable releases as newer than prerelease", async () => {
-			mockPackageJson.version = "0.1.0-beta.0";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "0.1.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("does not treat prerelease as newer than stable", async () => {
-			mockPackageJson.version = "0.1.0";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "0.1.0-beta.1" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("treats prerelease with additional segment as newer", async () => {
-			mockPackageJson.version = "1.2.3-beta";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.1" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("treats shorter prerelease as older than longer current prerelease", async () => {
-			mockPackageJson.version = "1.2.3-beta.1";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("compares numeric prerelease segments", async () => {
-			mockPackageJson.version = "1.2.3-beta.2";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.10" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("does not treat lower numeric prerelease as newer", async () => {
-			mockPackageJson.version = "1.2.3-beta.10";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.2" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("continues prerelease comparison after equal numeric segments", async () => {
-			mockPackageJson.version = "1.2.3-beta.2.omega";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.2.zeta" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("handles numeric current vs lexical latest prerelease segment", async () => {
-			mockPackageJson.version = "1.2.3-beta.2";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.alpha" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("handles lexical current vs numeric latest prerelease segment", async () => {
-			mockPackageJson.version = "1.2.3-beta.alpha";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.2" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("compares lexical prerelease segments by locale order", async () => {
-			mockPackageJson.version = "1.2.3-beta.omega";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.zeta" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-
-		it("does not treat lexically lower prerelease as newer", async () => {
-			mockPackageJson.version = "1.2.3-beta.zeta";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.2.3-beta.alpha" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(false);
-		});
-
-		it("handles invalid semver parts and build metadata safely", async () => {
-			mockPackageJson.version = "v1.x.0+build.1";
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "1.0.1+build.2" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-			expect(result.hasUpdate).toBe(true);
-		});
-	});
-
-	describe("checkForUpdates", () => {
-		it("logs debug details when package metadata cannot be parsed", async () => {
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return "{";
-				}
-				throw new Error("File not found");
-			});
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.currentVersion).toBe("0.0.0");
-			expect(globalThis.fetch).toHaveBeenCalled();
-			expect(logger.debug).toHaveBeenCalledWith(
-				"Failed to read current package version",
-				expect.objectContaining({ error: expect.any(String) }),
-			);
-		});
-
-		it("uses cache when check is recent", async () => {
-			const cacheData = {
-				lastCheck: Date.now() - 1000 * 60 * 60,
-				latestVersion: "5.0.0",
-				currentVersion: "4.12.0",
-			};
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				if (String(path).includes("update-check-cache.json")) {
-					return JSON.stringify(cacheData);
-				}
-				throw new Error("File not found");
-			});
-
-			const result = await checkForUpdates();
-
-			expect(globalThis.fetch).not.toHaveBeenCalled();
-			expect(result.hasUpdate).toBe(true);
-			expect(result.latestVersion).toBe("5.0.0");
-		});
-
-		it("logs debug details when cached update JSON is unreadable", async () => {
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				if (String(path).includes("update-check-cache.json")) {
-					return "{";
-				}
-				throw new Error("File not found");
-			});
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await checkForUpdates();
-
-			expect(logger.debug).toHaveBeenCalledWith(
-				"Failed to load update cache",
-				expect.objectContaining({ error: expect.any(String) }),
-			);
-			expect(globalThis.fetch).toHaveBeenCalled();
-		});
-
-		it.each(["EBUSY", "EPERM"] as const)(
-			"logs debug details when cache read fails on windows-style lock (%s)",
-			async (code) => {
-				vi.mocked(fs.existsSync).mockReturnValue(true);
-				vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-					if (String(path).includes("package.json")) {
-						return JSON.stringify(mockPackageJson);
-					}
-					if (String(path).includes("update-check-cache.json")) {
-						const error = new Error(`${code}: locked`) as NodeJS.ErrnoException;
-						error.code = code;
-						error.name = code;
-						throw error;
-					}
-					throw new Error("File not found");
-				});
-				vi.mocked(globalThis.fetch).mockResolvedValue({
-					ok: true,
-					json: async () => ({ version: "5.0.0" }),
-				} as Response);
-
-				await checkForUpdates();
-
-				expect(logger.debug).toHaveBeenCalledWith(
-					"Failed to load update cache",
-					expect.objectContaining({ error: expect.stringContaining(code) }),
-				);
-				expect(globalThis.fetch).toHaveBeenCalled();
-			},
-		);
-
-		it("handles cached null latestVersion without update", async () => {
-			const cacheData = {
-				lastCheck: Date.now() - 1000 * 60 * 60,
-				latestVersion: null,
-				currentVersion: "4.12.0",
-			};
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				if (String(path).includes("update-check-cache.json")) {
-					return JSON.stringify(cacheData);
-				}
-				throw new Error("File not found");
-			});
-
-			const result = await checkForUpdates();
-
-			expect(globalThis.fetch).not.toHaveBeenCalled();
-			expect(result.hasUpdate).toBe(false);
-			expect(result.latestVersion).toBeNull();
-		});
-
-		it("fetches when cache is expired", async () => {
-			const oldCache = {
-				lastCheck: Date.now() - 1000 * 60 * 60 * 25,
-				latestVersion: "4.11.0",
-				currentVersion: "4.12.0",
-			};
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				if (String(path).includes("update-check-cache.json")) {
-					return JSON.stringify(oldCache);
-				}
-				throw new Error("File not found");
-			});
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates();
-
-			expect(globalThis.fetch).toHaveBeenCalled();
-			expect(result.latestVersion).toBe("5.0.0");
-		});
-
-		it("fetches when force=true regardless of cache", async () => {
-			const freshCache = {
-				lastCheck: Date.now(),
-				latestVersion: "4.12.0",
-				currentVersion: "4.12.0",
-			};
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				if (String(path).includes("update-check-cache.json")) {
-					return JSON.stringify(freshCache);
-				}
-				throw new Error("File not found");
-			});
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "6.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(globalThis.fetch).toHaveBeenCalled();
-			expect(result.latestVersion).toBe("6.0.0");
-		});
-
-		it("handles fetch failure gracefully", async () => {
-			vi.mocked(globalThis.fetch).mockRejectedValue(new Error("Network error"));
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(false);
-			expect(result.latestVersion).toBe(null);
-		});
-
-		it("handles non-ok response", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: false,
-				status: 500,
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.hasUpdate).toBe(false);
-			expect(result.latestVersion).toBe(null);
-		});
-
-		it("saves cache after successful fetch", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await checkForUpdates(true);
-
-			expect(fs.writeFileSync).toHaveBeenCalled();
-			expect(fs.renameSync).toHaveBeenCalled();
-			const writeCall = vi.mocked(fs.writeFileSync).mock.calls[0];
-			const savedData = JSON.parse(writeCall[1] as string) as {
-				latestVersion: string;
-			};
-			expect(savedData.latestVersion).toBe("5.0.0");
-			expect(String(writeCall[0])).toContain("update-check-cache.json.");
-		});
-
-		it("creates cache directory if missing", async () => {
-			vi.mocked(fs.existsSync).mockReturnValue(false);
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await checkForUpdates(true);
-
-			expect(fs.mkdirSync).toHaveBeenCalledWith(expect.any(String), {
-				recursive: true,
-			});
-		});
-
-		it.each(["EBUSY", "EPERM"] as const)(
-			"retries cache writes when filesystem is transiently locked (%s)",
-			async (code) => {
-			let attempts = 0;
-			vi.mocked(fs.renameSync).mockClear();
-			vi.mocked(fs.renameSync).mockImplementation(() => {
-				attempts += 1;
-				if (attempts < 3) {
-					const error = new Error("busy") as NodeJS.ErrnoException;
-					error.code = code;
-					throw error;
-				}
-				return undefined;
-			});
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await checkForUpdates(true);
-
-			expect(fs.renameSync).toHaveBeenCalledTimes(3);
-			},
-		);
-
-		it("serializes concurrent cache writes through temp-file renames", async () => {
-			vi.mocked(fs.writeFileSync).mockClear();
-			vi.mocked(fs.renameSync).mockClear();
-			vi.mocked(globalThis.fetch)
-				.mockResolvedValueOnce({
-					ok: true,
-					json: async () => ({ version: "5.0.0" }),
-				} as Response)
-				.mockResolvedValueOnce({
-					ok: true,
-					json: async () => ({ version: "5.0.1" }),
-				} as Response);
-
-			await Promise.all([checkForUpdates(true), checkForUpdates(true)]);
-
-			expect(fs.renameSync).toHaveBeenCalledTimes(2);
-			const writeTargets = vi
-				.mocked(fs.writeFileSync)
-				.mock.calls.map((call) => String(call[0]));
-			expect(
-				writeTargets.every((target) =>
-					target.includes("update-check-cache.json."),
-				),
-			).toBe(true);
-		});
-
-		it("includes updateCommand in result", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			const result = await checkForUpdates(true);
-
-			expect(result.updateCommand).toContain("npm update -g");
-		});
-	});
-
-	describe("checkAndNotify", () => {
-		it("shows toast when update available", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const showToast = vi.fn().mockResolvedValue(undefined);
-
-			await checkAndNotify(showToast);
-
-			expect(showToast).toHaveBeenCalledWith(
-				expect.stringContaining("v5.0.0"),
-				"info"
-			);
-		});
-
-		it("does not show toast when no update", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "4.12.0" }),
-			} as Response);
-			const showToast = vi.fn().mockResolvedValue(undefined);
-
-			await checkAndNotify(showToast);
-
-			expect(showToast).not.toHaveBeenCalled();
-		});
-
-		it("handles missing showToast callback", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await expect(checkAndNotify()).resolves.toBeUndefined();
-		});
-
-		it("catches and ignores errors", async () => {
-			vi.mocked(globalThis.fetch).mockRejectedValue(new Error("Network error"));
-			const showToast = vi.fn().mockResolvedValue(undefined);
-
-			await expect(checkAndNotify(showToast)).resolves.toBeUndefined();
-			expect(showToast).not.toHaveBeenCalled();
-		});
-
-		it("catches toast callback errors", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const showToast = vi.fn().mockRejectedValue(new Error("toast failed"));
-
-			await expect(checkAndNotify(showToast)).resolves.toBeUndefined();
-			expect(showToast).toHaveBeenCalledTimes(1);
-			expect(showToast).toHaveBeenCalledWith(
-				expect.stringContaining("v5.0.0"),
-				"info",
-			);
-		});
-	});
-
-	describe("autoUpdateIfAvailable", () => {
-		function mockUpdateProcess(exitCode: number): EventEmitter & { kill: ReturnType<typeof vi.fn> } {
-			const child = new EventEmitter() as EventEmitter & {
-				kill: ReturnType<typeof vi.fn>;
-				unref: ReturnType<typeof vi.fn>;
-			};
-			child.kill = vi.fn();
-			child.unref = vi.fn();
-			const once = child.once.bind(child);
-			vi.spyOn(child, "once").mockImplementation((event, listener) => {
-				const result = once(event, listener);
-				if (event === "exit") {
-					void Promise.resolve().then(() => child.emit("exit", exitCode, null));
-				}
-				return result;
-			});
-			vi.mocked(childProcess.spawn).mockReturnValue(child as never);
-			return child;
-		}
-
-		function mockHangingUpdateProcess(): EventEmitter & { kill: ReturnType<typeof vi.fn>; unref: ReturnType<typeof vi.fn> } {
-			const child = new EventEmitter() as EventEmitter & {
-				kill: ReturnType<typeof vi.fn>;
-				unref: ReturnType<typeof vi.fn>;
-			};
-			child.kill = vi.fn();
-			child.unref = vi.fn();
-			vi.mocked(childProcess.spawn).mockReturnValue(child as never);
-			return child;
-		}
-
-		async function flushUpdateStartup(): Promise<void> {
-			for (let attempt = 0; attempt < 12; attempt += 1) {
-				await Promise.resolve();
-			}
-		}
-
-		it("defaults auto-update off in CI/test env and honors explicit opt-in", () => {
-			expect(isAutoUpdateEnabled({ CI: "true" })).toBe(false);
-			expect(isAutoUpdateEnabled({ VITEST: "true" })).toBe(false);
-			expect(isAutoUpdateEnabled({ CI: "true", CODEX_MULTI_AUTH_AUTO_UPDATE: "1" })).toBe(true);
-			expect(isAutoUpdateEnabled({ CODEX_MULTI_AUTH_AUTO_UPDATE: "0" })).toBe(false);
-		});
-
-		it("only treats installed node_modules packages as updateable by default", () => {
-			expect(
-				isUpdateablePackageInstall(
-					"/usr/local/lib/node_modules/codex-multi-auth",
-				),
-			).toBe(true);
-			expect(isUpdateablePackageInstall("C:/work/codex-multi-auth")).toBe(false);
-		});
-
-		it("resolves the package root above built dist modules", () => {
-			vi.mocked(fs.existsSync).mockImplementation((path: unknown) =>
-				String(path)
-					.replace(/\\/g, "/")
-					.endsWith("/node_modules/codex-multi-auth/package.json"),
-			);
-
-			expect(
-				resolvePackageRootFromModuleDir(
-					"C:/prefix/node_modules/codex-multi-auth/dist/lib",
-				).replace(/\\/g, "/"),
-			).toBe("C:/prefix/node_modules/codex-multi-auth");
-		});
-
-		it("skips without network work when disabled", async () => {
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "0" },
-				forceInstall: true,
-			});
-
-			expect(result.reason).toBe("disabled");
-			expect(globalThis.fetch).not.toHaveBeenCalled();
-			expect(childProcess.spawn).not.toHaveBeenCalled();
-		});
-
-		it("skips local worktrees unless forced", async () => {
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				packageRoot: "C:/work/codex-multi-auth",
-			});
-
-			expect(result.reason).toBe("not-updateable-install");
-			expect(globalThis.fetch).not.toHaveBeenCalled();
-			expect(childProcess.spawn).not.toHaveBeenCalled();
-		});
-
-		it("runs npm update when an update is available", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const child = mockUpdateProcess(0);
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			expect(result.reason).toBe("updated");
-			expect(result.updated).toBe(true);
-			expect(childProcess.spawn).toHaveBeenCalledWith(
-				"npm",
-				["update", "-g", "codex-multi-auth"],
-				expect.objectContaining({
-					env: expect.objectContaining({
-						CODEX_MULTI_AUTH_AUTO_UPDATE: "0",
-					}),
-					stdio: "ignore",
-				}),
-			);
-			expect(child.unref).toHaveBeenCalledTimes(1);
-			expect(fs.writeFileSync).toHaveBeenCalledWith(
-				expect.stringContaining("owner.json"),
-				expect.stringContaining(`"pid":${process.pid}`),
-				expect.objectContaining({ encoding: "utf8", mode: 0o600 }),
-			);
-			expect(fs.chmodSync).toHaveBeenCalledWith(
-				expect.stringContaining("owner.json"),
-				0o600,
-			);
-		});
-
-		it("runs Windows npm updates through cmd.exe", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			mockUpdateProcess(0);
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm.cmd",
-				platform: "win32",
-			});
-
-			expect(result.reason).toBe("updated");
-			expect(childProcess.spawn).toHaveBeenCalledWith(
-				expect.stringMatching(/(?:cmd|cmd\.exe)$/i),
-				[
-					"/d",
-					"/s",
-					"/c",
-					"npm.cmd update -g codex-multi-auth",
-				],
-				expect.objectContaining({
-					stdio: "ignore",
-					windowsHide: true,
-				}),
-			);
-		});
-
-		it("reports synchronous update spawn failures without throwing", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			vi.mocked(childProcess.spawn).mockImplementation(() => {
-				throw new Error("spawn EINVAL");
-			});
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			expect(result.reason).toBe("update-failed");
-			expect(result.error).toBe("spawn EINVAL");
-		});
-
-		it("reports asynchronous update spawn errors without throwing", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			vi.mocked(childProcess.spawn).mockImplementation(() => {
-				const child = new EventEmitter() as EventEmitter & {
-					kill: ReturnType<typeof vi.fn>;
-				};
-				child.kill = vi.fn();
-				void Promise.resolve().then(() => child.emit("error", new Error("spawn EACCES")));
-				return child as never;
-			});
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			expect(result.reason).toBe("update-failed");
-			expect(result.error).toBe("spawn EACCES");
-			expect(logger.warn).toHaveBeenCalledWith(
-				"Auto-update failed",
-				expect.objectContaining({ error: "spawn EACCES" }),
-			);
-		});
-
-		it("times out hung npm update processes and kills the child", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const child = mockHangingUpdateProcess();
-
-			const resultPromise = autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-				timeoutMs: 25,
-			});
-			await flushUpdateStartup();
-			expect(childProcess.spawn).toHaveBeenCalledTimes(1);
-			await vi.advanceTimersByTimeAsync(25);
-			const result = await resultPromise;
-
-			expect(child.kill).toHaveBeenCalledTimes(1);
-			expect(result.reason).toBe("update-failed");
-			expect(result.exitCode).toBeNull();
-			expect(result.error).toContain("Auto-update timed out after 25ms");
-		});
-
-		it("uses taskkill to clean up hung Windows update process trees", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const child = new EventEmitter() as EventEmitter & {
-				kill: ReturnType<typeof vi.fn>;
-				unref: ReturnType<typeof vi.fn>;
-				pid: number;
-			};
-			child.kill = vi.fn();
-			child.unref = vi.fn();
-			child.pid = 1234;
-			const killer = new EventEmitter();
-			vi.mocked(childProcess.spawn)
-				.mockReturnValueOnce(child as never)
-				.mockReturnValueOnce(killer as never);
-
-			const resultPromise = autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm.cmd",
-				platform: "win32",
-				timeoutMs: 25,
-			});
-			await flushUpdateStartup();
-			await vi.advanceTimersByTimeAsync(25);
-			expect(childProcess.spawn).toHaveBeenLastCalledWith(
-				"cmd.exe",
-				[
-					"/d",
-					"/s",
-					"/c",
-					"taskkill /PID 1234 /T /F",
-				],
-				expect.objectContaining({ stdio: "ignore", windowsHide: true }),
-			);
-			killer.emit("exit", 0, null);
-			const result = await resultPromise;
-
-			expect(child.kill).not.toHaveBeenCalled();
-			expect(result.reason).toBe("update-failed");
-		});
-
-		it("falls back to killing the child when Windows taskkill fails", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			const child = new EventEmitter() as EventEmitter & {
-				kill: ReturnType<typeof vi.fn>;
-				unref: ReturnType<typeof vi.fn>;
-				pid: number;
-			};
-			child.kill = vi.fn();
-			child.unref = vi.fn();
-			child.pid = 1234;
-			const killer = new EventEmitter();
-			vi.mocked(childProcess.spawn)
-				.mockReturnValueOnce(child as never)
-				.mockReturnValueOnce(killer as never);
-
-			const resultPromise = autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm.cmd",
-				platform: "win32",
-				timeoutMs: 25,
-			});
-			await flushUpdateStartup();
-			await vi.advanceTimersByTimeAsync(25);
-			killer.emit("exit", 1, null);
-			const result = await resultPromise;
-
-			expect(child.kill).toHaveBeenCalledTimes(1);
-			expect(result.reason).toBe("update-failed");
-		});
-
-		it("passes a least-privilege environment to npm update", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			mockUpdateProcess(0);
-
-			await autoUpdateIfAvailable({
-				env: {
-					CODEX_MULTI_AUTH_AUTO_UPDATE: "1",
-					GITHUB_TOKEN: "ghp_secret",
-					NODE_AUTH_TOKEN: "npm_secret",
-					OPENAI_API_KEY: "sk-secret",
-					PATH: "C:/tools",
-					npm_config_registry: "https://registry.npmjs.org/",
-					npm_config__authToken: "npm_secret",
-				},
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			const spawnEnv = vi.mocked(childProcess.spawn).mock.calls[0]?.[2]?.env as NodeJS.ProcessEnv;
-			expect(spawnEnv.CODEX_MULTI_AUTH_AUTO_UPDATE).toBe("0");
-			expect(spawnEnv.PATH).toBe("C:/tools");
-			expect(spawnEnv.npm_config_registry).toBe("https://registry.npmjs.org/");
-			expect(spawnEnv.GITHUB_TOKEN).toBeUndefined();
-			expect(spawnEnv.NODE_AUTH_TOKEN).toBeUndefined();
-			expect(spawnEnv.OPENAI_API_KEY).toBeUndefined();
-			expect(spawnEnv.npm_config__authToken).toBeUndefined();
-		});
-
-		it("skips concurrent npm updates when another process holds the lock", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			let lockHeld = false;
-			let child: (EventEmitter & { kill: ReturnType<typeof vi.fn> }) | null = null;
-			vi.mocked(fs.existsSync).mockImplementation((path: unknown) =>
-				!String(path).includes("auto-update.lock"),
-			);
-			vi.mocked(fs.mkdirSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("auto-update.lock")) {
-					if (lockHeld) {
-						const error = new Error("lock exists") as NodeJS.ErrnoException;
-						error.code = "EEXIST";
-						throw error;
-					}
-					lockHeld = true;
-				}
-				return undefined as never;
-			});
-			vi.mocked(fs.rmdirSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("auto-update.lock")) {
-					lockHeld = false;
-				}
-			});
-			vi.mocked(childProcess.spawn).mockImplementation(() => {
-				child = new EventEmitter() as EventEmitter & { kill: ReturnType<typeof vi.fn> };
-				child.kill = vi.fn();
-				return child as never;
-			});
-
-			const first = autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-			const second = autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-			await flushUpdateStartup();
-			expect(childProcess.spawn).toHaveBeenCalledTimes(1);
-			child?.emit("exit", 0, null);
-
-			const results = await Promise.all([first, second]);
-
-			expect(results.map((result) => result.reason).sort()).toEqual([
-				"already-running",
-				"updated",
-			]);
-		});
-
-		it("does not steal stale-looking locks while the owner pid is alive", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			vi.mocked(fs.mkdirSync).mockImplementation((path: unknown) => {
-				if (String(path).includes("auto-update.lock")) {
-					const error = new Error("lock exists") as NodeJS.ErrnoException;
-					error.code = "EEXIST";
-					throw error;
-				}
-				return undefined as never;
-			});
-			vi.mocked(fs.statSync).mockReturnValue({
-				mtimeMs: Date.now() - 20 * 60_000,
-			} as never);
-			vi.mocked(fs.readFileSync).mockImplementation((path: unknown) => {
-				const value = String(path);
-				if (value.includes("owner.json")) {
-					return JSON.stringify({ pid: process.pid });
-				}
-				if (value.includes("package.json")) {
-					return JSON.stringify(mockPackageJson);
-				}
-				throw new Error("File not found");
-			});
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			expect(result.reason).toBe("already-running");
-			expect(childProcess.spawn).not.toHaveBeenCalled();
-			expect(fs.rmdirSync).not.toHaveBeenCalled();
-		});
-
-		it("reports npm update failures without throwing", async () => {
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-			mockUpdateProcess(1);
-
-			const result = await autoUpdateIfAvailable({
-				env: { CODEX_MULTI_AUTH_AUTO_UPDATE: "1" },
-				forceCheck: true,
-				forceInstall: true,
-				npmCommand: "npm",
-				platform: "linux",
-			});
-
-			expect(result.reason).toBe("update-failed");
-			expect(result.updated).toBe(false);
-			expect(result.exitCode).toBe(1);
-			expect(logger.warn).toHaveBeenCalledWith(
-				"Auto-update failed",
-				expect.objectContaining({ error: expect.any(String) }),
-			);
-		});
-	});
-
-	describe("clearUpdateCache", () => {
-		it("writes empty object when cache exists", () => {
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.writeFileSync).mockClear();
-			vi.mocked(fs.renameSync).mockClear();
-
-			clearUpdateCache();
-
-			return Promise.resolve().then(() => {
-				return Promise.resolve().then(() => {
-					expect(fs.writeFileSync).toHaveBeenCalledWith(
-						expect.stringContaining("update-check-cache.json."),
-						"{}",
-						"utf8"
-					);
-					expect(fs.renameSync).toHaveBeenCalled();
-				});
-			});
-		});
-
-		it("does nothing when cache does not exist", () => {
-			vi.mocked(fs.existsSync).mockReturnValue(false);
-			vi.mocked(fs.writeFileSync).mockClear();
-
-			clearUpdateCache();
-
-			expect(fs.writeFileSync).not.toHaveBeenCalled();
-		});
-
-		it("keeps clearUpdateCache ordered after the latest save", async () => {
-			vi.mocked(fs.existsSync).mockReturnValue(true);
-			vi.mocked(fs.writeFileSync).mockClear();
-			vi.mocked(fs.renameSync).mockClear();
-			vi.mocked(globalThis.fetch).mockResolvedValue({
-				ok: true,
-				json: async () => ({ version: "5.0.0" }),
-			} as Response);
-
-			await checkForUpdates(true);
-			clearUpdateCache();
-			await Promise.resolve();
-			await Promise.resolve();
-
-			const writes = vi.mocked(fs.writeFileSync).mock.calls;
-			expect(writes).toHaveLength(2);
-			expect(writes.at(-1)?.[1]).toBe("{}");
-		});
-	});
-});
