+++ test/codex-bin-wrapper.test.ts
@@ -3894,26 +3894,26 @@ describe("codex bin wrapper", () => {
 		expect(result.stdout).toContain("FORWARDED:auth status");
 	});
 
-	it("skips startup auto-update loading when bypass is set", () => {
+	it("skips startup update-notice loading when bypass is set", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
-		const markerPath = join(fixtureRoot, "auto-update-loaded.txt");
+		const markerPath = join(fixtureRoot, "update-notice-loaded.txt");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
 				'import { writeFileSync } from "node:fs";',
-				"writeFileSync(process.env.CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER, 'loaded', 'utf8');",
-				"export async function autoUpdateIfAvailable() {",
-				"\treturn { updated: false };",
+				"writeFileSync(process.env.CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER, 'loaded', 'utf8');",
+				"export async function checkForUpdates() {",
+				"\treturn { hasUpdate: false };",
 				"}",
 			].join("\n"),
 			"utf8",
 		);
 
 		const result = runWrapper(fixtureRoot, ["--version"], {
-			CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER: markerPath,
+			CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER: markerPath,
 			CODEX_MULTI_AUTH_BYPASS: "1",
 			CODEX_MULTI_AUTH_REAL_CODEX_BIN: fakeBin,
 		});
@@ -3929,26 +3929,26 @@ describe("codex bin wrapper", () => {
 		["long help", ["--help"]],
 		["short help", ["-h"]],
 		["combined help/version", ["--help", "--version"]],
-	] as const)("skips startup auto-update loading for pure %s commands", (_label, args) => {
+	] as const)("skips startup update-notice loading for pure %s commands", (_label, args) => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
-		const markerPath = join(fixtureRoot, "auto-update-loaded.txt");
+		const markerPath = join(fixtureRoot, "update-notice-loaded.txt");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
 				'import { writeFileSync } from "node:fs";',
-				"writeFileSync(process.env.CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER, 'loaded', 'utf8');",
-				"export async function autoUpdateIfAvailable() {",
-				"\treturn { updated: false };",
+				"writeFileSync(process.env.CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER, 'loaded', 'utf8');",
+				"export async function checkForUpdates() {",
+				"\treturn { hasUpdate: false };",
 				"}",
 			].join("\n"),
 			"utf8",
 		);
 
 		const result = runWrapper(fixtureRoot, [...args], {
-			CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER: markerPath,
+			CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER: markerPath,
 			CODEX_MULTI_AUTH_REAL_CODEX_BIN: fakeBin,
 		});
 
@@ -3957,18 +3957,18 @@ describe("codex bin wrapper", () => {
 		expect(existsSync(markerPath)).toBe(false);
 	});
 
-	it("skips startup auto-update loading for local auth commands", () => {
+	it("skips startup update-notice loading for local auth commands", () => {
 		const fixtureRoot = createWrapperFixture();
 		const distLibDir = join(fixtureRoot, "dist", "lib");
-		const markerPath = join(fixtureRoot, "auto-update-loaded.txt");
+		const markerPath = join(fixtureRoot, "update-notice-loaded.txt");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
 				'import { writeFileSync } from "node:fs";',
-				"writeFileSync(process.env.CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER, 'loaded', 'utf8');",
-				"export async function autoUpdateIfAvailable() {",
-				"\treturn { updated: false };",
+				"writeFileSync(process.env.CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER, 'loaded', 'utf8');",
+				"export async function checkForUpdates() {",
+				"\treturn { hasUpdate: false };",
 				"}",
 			].join("\n"),
 			"utf8",
@@ -3985,7 +3985,7 @@ describe("codex bin wrapper", () => {
 		);
 
 		const result = runWrapper(fixtureRoot, ["auth", "status"], {
-			CODEX_MULTI_AUTH_AUTO_UPDATE_MARKER: markerPath,
+			CODEX_MULTI_AUTH_UPDATE_NOTICE_MARKER: markerPath,
 		});
 
 		expect(result.status).toBe(0);
@@ -3993,7 +3993,7 @@ describe("codex bin wrapper", () => {
 		expect(existsSync(markerPath)).toBe(false);
 	});
 
-	it("ignores missing startup auto-update checker builds", () => {
+	it("ignores missing startup update-notice builds", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 
@@ -4003,36 +4003,38 @@ describe("codex bin wrapper", () => {
 
 		expect(result.status).toBe(0);
 		expect(result.stdout).toContain("FORWARDED:exec status");
-		expect(result.stderr).not.toContain("auto-update skipped");
+		expect(result.stderr).not.toContain("update notice skipped");
 		expect(result.stderr).not.toContain(
-			"codex-multi-auth: auto-update found 9.9.9; starting npm update -g codex-multi-auth.",
+			"codex-multi-auth update available: v9.9.9",
 		);
 		expect(result.stderr).not.toContain(
-			"codex-multi-auth: auto-updated to 9.9.9. New sessions will use the latest package.",
+			"npm install -g codex-multi-auth@latest",
 		);
 	});
 
-	it("logs startup auto-update progress and successful updates in debug mode", () => {
+	it("logs startup update notices with the manual install command in debug mode", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
-		const optionsPath = join(fixtureRoot, "auto-update-options.json");
+		const optionsPath = join(fixtureRoot, "update-notice-options.json");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
 				'import { writeFileSync } from "node:fs";',
-				"export async function autoUpdateIfAvailable(options) {",
-				"\twriteFileSync(process.env.CODEX_MULTI_AUTH_AUTO_UPDATE_OPTIONS, JSON.stringify({ fetchTimeoutMs: options.fetchTimeoutMs, timeoutMs: options.timeoutMs }), 'utf8');",
-				"\toptions?.onUpdateStart?.({ latestVersion: '9.9.9' });",
-				"\treturn { updated: true, latestVersion: '9.9.9' };",
+				"export async function checkForUpdates(force, fetchTimeoutMs) {",
+				"\twriteFileSync(process.env.CODEX_MULTI_AUTH_UPDATE_NOTICE_OPTIONS, JSON.stringify({ force, fetchTimeoutMs }), 'utf8');",
+				"\treturn { hasUpdate: true, currentVersion: '1.0.0', latestVersion: '9.9.9', updateCommand: 'npm install -g codex-multi-auth@latest' };",
+				"}",
+				"export function formatManualUpdateNotice(result) {",
+				"\treturn `codex-multi-auth update available: v${result.latestVersion}; current: v${result.currentVersion}; run: ${result.updateCommand}`;",
 				"}",
 			].join("\n"),
 			"utf8",
 		);
 
 		const result = runWrapper(fixtureRoot, ["exec", "status"], {
-			CODEX_MULTI_AUTH_AUTO_UPDATE_OPTIONS: optionsPath,
+			CODEX_MULTI_AUTH_UPDATE_NOTICE_OPTIONS: optionsPath,
 			CODEX_MULTI_AUTH_DEBUG: "1",
 			CODEX_MULTI_AUTH_REAL_CODEX_BIN: fakeBin,
 		});
@@ -4040,28 +4042,25 @@ describe("codex bin wrapper", () => {
 		expect(result.status).toBe(0);
 		expect(result.stdout).toContain("FORWARDED:exec status");
 		expect(JSON.parse(readFileSync(optionsPath, "utf8"))).toEqual({
-			fetchTimeoutMs: 1200,
-			timeoutMs: 1800,
+			force: false,
+			fetchTimeoutMs: 2400,
 		});
 		expect(result.stderr).toContain(
-			"codex-multi-auth: auto-update found 9.9.9; starting npm update -g codex-multi-auth. Startup will continue if it exceeds 3000ms.",
+			"codex-multi-auth update available: v9.9.9; current: v1.0.0; run: npm install -g codex-multi-auth@latest",
 		);
-		expect(result.stderr).toContain(
-			"codex-multi-auth: auto-updated to 9.9.9. New sessions will use the latest package.",
-		);
+		expect(result.stderr).not.toContain("npm update -g");
 	});
 
-	it("suppresses startup auto-update progress in captured non-TTY output", () => {
+	it("suppresses startup update notices in captured non-TTY output", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
-				"export async function autoUpdateIfAvailable(options) {",
-				"\toptions?.onUpdateStart?.({ latestVersion: '9.9.9' });",
-				"\treturn { updated: true, latestVersion: '9.9.9' };",
+				"export async function checkForUpdates() {",
+				"\treturn { hasUpdate: true, currentVersion: '1.0.0', latestVersion: '9.9.9', updateCommand: 'npm install -g codex-multi-auth@latest' };",
 				"}",
 			].join("\n"),
 			"utf8",
@@ -4073,19 +4072,19 @@ describe("codex bin wrapper", () => {
 
 		expect(result.status).toBe(0);
 		expect(result.stdout).toContain("FORWARDED:exec status");
-		expect(result.stderr).not.toContain("auto-update found 9.9.9");
-		expect(result.stderr).not.toContain("auto-updated to 9.9.9");
+		expect(result.stderr).not.toContain("update available");
+		expect(result.stderr).not.toContain("npm install -g codex-multi-auth@latest");
 	});
 
-	it("suppresses startup auto-update failures unless debug logging is enabled", () => {
+	it("suppresses startup update-notice failures unless debug logging is enabled", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
-				"export async function autoUpdateIfAvailable() {",
+				"export async function checkForUpdates() {",
 				"\tthrow new Error('registry unavailable');",
 				"}",
 			].join("\n"),
@@ -4106,19 +4105,19 @@ describe("codex bin wrapper", () => {
 		expect(debugResult.status).toBe(0);
 		expect(debugResult.stdout).toContain("FORWARDED:exec status");
 		expect(debugResult.stderr).toContain(
-			"codex-multi-auth: auto-update skipped: registry unavailable",
+			"codex-multi-auth: update notice skipped: registry unavailable",
 		);
 	});
 
-	it("continues forwarded startup within the stable auto-update startup budget", () => {
+	it("continues forwarded startup within the stable update-notice startup budget", () => {
 		const fixtureRoot = createWrapperFixture();
 		const fakeBin = createFakeCodexBin(fixtureRoot);
 		const distLibDir = join(fixtureRoot, "dist", "lib");
 		mkdirSync(distLibDir, { recursive: true });
 		writeFileSync(
-			join(distLibDir, "auto-update-checker.js"),
+			join(distLibDir, "update-notice.js"),
 			[
-				"export async function autoUpdateIfAvailable() {",
+				"export async function checkForUpdates() {",
 				"\treturn new Promise(() => undefined);",
 				"}",
 			].join("\n"),
@@ -4128,13 +4127,13 @@ describe("codex bin wrapper", () => {
 		const startedAt = Date.now();
 		const result = runWrapper(fixtureRoot, ["exec", "status"], {
 			CODEX_MULTI_AUTH_REAL_CODEX_BIN: fakeBin,
-			CODEX_MULTI_AUTH_AUTO_UPDATE_STARTUP_BUDGET_MS: "25",
+			CODEX_MULTI_AUTH_TEST_STARTUP_UPDATE_NOTICE_BUDGET_MS: "25",
 		});
 		const elapsedMs = Date.now() - startedAt;
 
 		expect(result.status).toBe(0);
 		expect(result.stdout).toContain("FORWARDED:exec status");
-		expect(result.stderr).not.toContain("auto-update skipped");
+		expect(result.stderr).not.toContain("update notice skipped");
 		expect(elapsedMs).toBeLessThan(2_000);
 	});
